refactored the code: separated the classes based on use-cases - elasticsearch utils - json utils - xmls utils - gpt utils
