+++ push_summary_to_es.py
@@ -1,202 +1,10 @@
-import re
-from elasticsearch import Elasticsearch
-import time
 from datetime import datetime, timedelta
 from loguru import logger
-import xml.etree.ElementTree as ET
-import os
-from dotenv import load_dotenv
-import warnings
-import pytz
 import tqdm
 
-from src.config import ES_CLOUD_ID, ES_USERNAME, ES_PASSWORD, ES_INDEX, ES_DATA_FETCH_SIZE
-
-warnings.filterwarnings("ignore")
-load_dotenv()
-
-
-class ElasticSearchClient:
-    def __init__(self, es_cloud_id, es_username, es_password, es_data_fetch_size=ES_DATA_FETCH_SIZE) -> None:
-        self._es_cloud_id = es_cloud_id
-        self._es_username = es_username
-        self._es_password = es_password
-        self._es_data_fetch_size = es_data_fetch_size
-        self._es_client = Elasticsearch(
-            cloud_id=self._es_cloud_id,
-            http_auth=(self._es_username, self._es_password),
-        )
-
-    def fetch_data_with_empty_summary(self, es_index, url=None, start_date_str=None, current_date_str=None):
-        logger.info(f"connecting ElasticSearch to fetch the docs with summary ... ")
-        output_list = []
-        start_time = time.time()
-
-        if self._es_client.ping():
-            logger.success("connected to the ElasticSearch")
-
-            if url and start_date_str and current_date_str:
-                logger.info(f"Url: {url}, Start Date: {start_date_str}, Current Date: {current_date_str}")
-                query = {
-                    "query": {
-                        "bool": {
-                            "must": [
-                                {
-                                    "range": {
-                                        "created_at": {
-                                            "gte": f"{start_date_str}T00:00:00.000Z",
-                                            "lte": f"{current_date_str}T23:59:59.999Z"
-                                        }
-                                    }
-                                },
-                                {
-                                    "prefix": {
-                                        "domain.keyword": str(url)
-                                    }
-                                }
-                            ],
-                            "must_not": {
-                                "exists": {
-                                    "field": "summary"
-                                }
-                            }
-                        }
-                    }
-                }
-            elif url and not start_date_str and not current_date_str:
-                logger.info(f"Url: {url}, Start Date: {start_date_str}, Current Date: {current_date_str}")
-                query = {
-                    "query": {
-                        "bool": {
-                            "must": [
-                                {
-                                    "prefix": {
-                                        "domain.keyword": str(url)
-                                    }
-                                }
-                            ],
-                            "must_not": {
-                                "exists": {
-                                    "field": "summary"
-                                }
-                            }
-                        }
-                    }
-                }
-            elif not url and start_date_str and current_date_str:
-                logger.info(f"Url: {url}, Start Date: {start_date_str}, Current Date: {current_date_str}")
-                query = {
-                    "query": {
-                        "bool": {
-                            "must": [
-                                {
-                                    "range": {
-                                        "created_at": {
-                                            "gte": f"{start_date_str}T00:00:00.000Z",
-                                            "lte": f"{current_date_str}T23:59:59.999Z"
-                                        }
-                                    }
-                                }
-                            ],
-                            "must_not": {
-                                "exists": {
-                                    "field": "summary"
-                                }
-                            }
-                        }
-                    }
-                }
-            else:
-                logger.info(f"Url: {url}, Start Date: {start_date_str}, Current Date: {current_date_str}")
-                query = {
-                    "query": {
-                        "bool": {
-                            "must_not": {
-                                "exists": {
-                                    "field": "summary"
-                                }
-                            }
-                        }
-                    }
-                }
-
-            # Initialize the scroll
-            scroll_response = self._es_client.search(index=es_index, body=query, size=self._es_data_fetch_size,
-                                                     scroll='1m')
-            scroll_id = scroll_response['_scroll_id']
-            results = scroll_response['hits']['hits']
-
-            # Dump the documents into the json file
-            logger.info(f"started dumping of {es_index} data in json...")
-            while len(results) > 0:
-                # Save the current batch of results
-                for result in results:
-                    output_list.append(result)
-
-                # Fetch the next batch of results
-                scroll_response = self._es_client.scroll(scroll_id=scroll_id, scroll='1m')
-                scroll_id = scroll_response['_scroll_id']
-                results = scroll_response['hits']['hits']
-
-            logger.info(
-                f"Dumping of {es_index} data in json has completed and has taken {time.time() - start_time:.2f} seconds.")
-
-            return output_list
-        else:
-            logger.warning('Could not connect to Elasticsearch')
-            return output_list
-
-    @property
-    def es_client(self):
-        return self._es_client
-
-
-class XMLReader:
-    def __init__(self) -> None:
-        self.month_dict = {
-            1: "Jan", 2: "Feb", 3: "March", 4: "April", 5: "May", 6: "June",
-            7: "July", 8: "Aug", 9: "Sept", 10: "Oct", 11: "Nov", 12: "Dec"
-        }
-
-    def get_id(self, id):
-        return str(id).split("-")[-1]
-
-    def clean_title(self, xml_name):
-        special_characters = ['/', ':', '@', '#', '$', '*', '&', '<', '>', '\\', '?']
-        xml_name = re.sub(r'[^A-Za-z0-9]+', '-', xml_name)
-        for sc in special_characters:
-            xml_name = xml_name.replace(sc, "-")
-        return xml_name
-
-    def get_xml_summary(self, data, dev_name):
-        number = self.get_id(data["_source"]["id"])
-        title = data["_source"]["title"]
-        xml_name = self.clean_title(title)
-        published_at = datetime.strptime(data['_source']['created_at'], '%Y-%m-%dT%H:%M:%S.%fZ')
-        published_at = pytz.UTC.localize(published_at)
-        month_name = self.month_dict[int(published_at.month)]
-        str_month_year = f"{month_name}_{int(published_at.year)}"
-        current_directory = os.getcwd()
-        file_path = f"static/{dev_name}/{str_month_year}/{number}_{xml_name}.xml"
-        full_path = os.path.join(current_directory, file_path)
-
-        try:
-            if os.path.exists(full_path):
-                namespaces = {'atom': 'http://www.w3.org/2005/Atom'}
-                tree = ET.parse(full_path)
-                root = tree.getroot()
-                summ_list = root.findall(".//atom:entry/atom:summary", namespaces)
-                if summ_list:
-                    summ = "\n".join([summ.text for summ in summ_list])
-                    return summ, f"Summary text found: {full_path}"
-                else:
-                    return None, f"No summary found: {full_path}"
-            else:
-                return None, f"No xml file found: {full_path}"
-        except Exception as e:
-            ex_message = f"Error: {e} at file path: {full_path}"
-            logger.error(ex_message)
-            return None, ex_message
+from src.config import ES_INDEX
+from src.elasticsearch_utils import ElasticSearchClient
+from src.xml_utils import XMLReader
 
 
 if __name__ == "__main__":
@@ -204,8 +12,7 @@ if __name__ == "__main__":
     APPLY_DATE_RANGE = False
 
     xml_reader = XMLReader()
-    elastic_search = ElasticSearchClient(es_cloud_id=ES_CLOUD_ID, es_username=ES_USERNAME,
-                                         es_password=ES_PASSWORD)
+    elastic_search = ElasticSearchClient()
 
     dev_urls = [
         "https://lists.linuxfoundation.org/pipermail/bitcoin-dev/",
