+++ src/config.py
@@ -1,7 +1,19 @@
-import tiktoken
 import os
+import openai
+import warnings
+
+import tiktoken
 from dotenv import load_dotenv
+
 load_dotenv()
+
+warnings.filterwarnings("ignore")
+
+OPENAI_ORG_KEY = os.getenv("OPENAI_ORG_KEY")
+OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
+
+openai.organization = OPENAI_ORG_KEY
+openai.api_key = OPENAI_API_KEY
 
 TOKENIZER = tiktoken.get_encoding("cl100k_base")
 
@@ -12,7 +24,6 @@ COMPLETION_MODEL = "text-davinci-003"  # "text-ada-001"
 
 # used in generation of xmls production, homepage, newsletter etc.
 CHAT_COMPLETION_MODEL = "gpt-4-1106-preview"  # "gpt-3.5-turbo", "gpt-4"
-OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
 
 ES_CLOUD_ID = os.getenv("ES_CLOUD_ID")
 ES_USERNAME = os.getenv("ES_USERNAME")
