--- src/logger.py
@@ -1,28 +1,0 @@
-import logging
-import sys
-
-import warnings
-warnings.filterwarnings("ignore")
-
-LOGGER = logging.getLogger()
-FORMATTER = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-
-
-def setup_logger(out_file=None, stderr=True, stderr_level=logging.INFO, file_level=logging.DEBUG):
-    LOGGER.handlers = []
-    LOGGER.setLevel(min(stderr_level, file_level))
-
-    if stderr:
-        handler = logging.StreamHandler(sys.stderr)
-        handler.setFormatter(FORMATTER)
-        handler.setLevel(stderr_level)
-        LOGGER.addHandler(handler)
-
-    if out_file is not None:
-        handler = logging.FileHandler(out_file)
-        handler.setFormatter(FORMATTER)
-        handler.setLevel(file_level)
-        LOGGER.addHandler(handler)
-
-    LOGGER.info("logger set up")
-    return LOGGER
